+++ src/pages/posts/Posts.module.css
@@ -10,14 +10,6 @@
     width: 592px;
 }
 
-.post_info {
-    margin-top: 20px;
-}
-
-.post_info_text {
-    font-size: 24px;
-}
-
 /* 버튼 */
 .go_upload {
     margin-top: 2vh;
