+++ src/pages/posts/index.jsx
@@ -1,0 +1,3 @@
+import Posts from './Posts';
+
+export default Posts;
