+++ src/pages/register/Register.module.css
@@ -1,0 +1,3 @@
+.main {
+    margin-bottom: 20vh;
+}
